backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sabuhi Ibrahimov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.6"
django-rest-framework = "^0.1.0"
celery = "^5.3.4"
python-jose = "^3.3.0"
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"
# psycopg2 = "^2.9.9"
orjson = "^3.9.8"
pillow = "^10.0.1"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
